Real fix for RT #128156
Check all dependencies first before starting to load any of them.
More importantly: unconditionally check all dependencies before trying
to load a precomp file. Previously we did not check a dependency if it
has already been loaded.

Consider: A depending on C and B. B depending on C.  If C is outdated,
we already notice that trying to load A. We recompile C and load it.
Then when it's time to check B's dependencies, we don't check C because
it's already been loaded and therefore not notice that B is outdated, too.

The previous fix only hid the problem by having us delete B's precomp file
as a reverse dependency of C upon recompilation. But that wouldn't work if
B was in a different repository.