Remove .rev-deps cache
.rev-deps files were a good idea at the time for tracking which reverse-
dependencies we have to recompile after updating a module. But that was when
we only ever used a single precomp store at a time, so everything of interest
would be found in the same store.

Now a reverse dependency may be in a store further up the chain and .rev-deps
files are possibly incomplete. They are also a problem for packaging modules
as they are changed on installation of other modules. Removing them also clears
the way for fixing the PrecompilationStore abstraction.