Simplify REPL error reporting
It occurred to me that using the standard error reporting in the REPL
may be a bit overkill:

> $number = 5;
===SORRY!=== Error while compiling <unknown file>
Variable '$number' is not declared
at <unknown file>:1
------> <BOL><HERE>$number = 5;

For one, the "unknown file" is useless information.  The "at <unknown file>:1"
also.  This patch reduces the reporting to just the line:

Variable '$number' is not declared

which may be a bit too sparse.  But I gladly leave this up to someone
else to get the line:

------> <BOL><HERE>$number = 5;

back in again.

Anyways, please revert if this is too simplifying.