Make the List constructore more memory friendly
The difference in CPU usage is within noise levels.  However, it does
not allocate an IterationBuffer if there are no Slips in the parameter
list (which is usually the case when creating a List, I would argue).

This could get significantly faster if we had nqp::listhastype(list,Type)
op that would return 1 if the list had any element of the given type.