Revert handling of 'No' chars in Str.Int
The feature was added as a fix for RT#127866, however, Str.Numeric and
especially val() remained unable to handle 'No' characters.

After conversation with TimToady
(http://irclog.perlgeek.de/perl6/2016-07-04#i_12782986), it was ruled
that handling of 'No' characters should be explicitly requested
by the programmer with unival() and so this commit undoes such handling
in Str.Int case to maintain consistency.