Fix BEGIN time EVAL screwing up MAIN_CTX
When compiling a source file, we store the compilation unit's main context
for later merging with the outer compilation unit's globals. This storing
is done by the compiler calling the ctxsave method of an object stored in
a contextual - $*CTX_SAVE - when it finished the compilation unit. The
ctxsave method ensures that it's called only once.

Now EVALs are their own compilation unit. When running an EVAL at BEGIN time
of a currently compiling source file, the compiler finds a $*CTX_SAVE and
uses it to store the EVAL's main context. The source file's context won't be
stored anymore after that.

Fix by preventing the compiler from seeing $*CTX_SAVE when compiling an EVAL.
Fixes RT #128457