Make circumfix:[ ]> about 2.5x as fast
Turns out that the Iterable:D candidate was basically doing all of the
work, because as soon as we have any , within [], it creates a List
using infix:<,>.  Since this is an Iterable, it would be caught by the
Iterable:D candidate, which then assumed it was something strange and
started the whole iterator process.  Instead, we now check whether it
is a non-lazy List (which is usually the case), we simply copy all of
the elements over into containers and return the resulting Array.