Re-imagine List.from-slurpy-flat
It struck me that if a sub has a slurpy array, and the values specified
are all simple scalars without any additional Slippy/Iterable functionality,
that the $!future created (which is the $!todo of the List/Array to be)
is in fact already what $!reified will be in the end.  So in that case
we can just put the $!future into $!reified, and we're done:

  sub a(*@a) { }; a(1,2,3,4);

sees an improvement of 30% compared to a(), and 3 allocations less.

This spectests clean.  But this is the area of dragons, so there may
be some ecosystem fallout.  In that case, just revert this commit.