Don't include repositories in precomp dependencies
Just because we loaded a custom repo implementation like
CompUnit::Repository::Staging during precompilation of a module doesn't
mean that the module depends on this repo implementation at runtime.

This fixes a bootstrap issue: the Staging repo is for precompiling a
module into a staging area which later on gets copied into the final
place. The staging area's path however must not end up in the precomp
files and should instead be replaced by the repo's logical name if
available.

When installing CompUnit::Repository::Staging itself however, this
mechanism is not yet available.