Make "my @a is List = 1,2,3" work
In many cases, one wants to have Positional access on a List, rather
than an Array, because once initialized, the List cannot change.
This syntax, which was supposed to work anyway as similar syntax for
"my %h is Map = a => 42" also works, has the advantage that it is
cheaper to initialize, as there is no need to create containers for
the values (since we do not intend to change it anyway).  This saves
at least one allocation per reified element, so easier on the GC and
more efficient on the CPU as well.