Fix merge global symbols failures when changing repositories
Given:
use lib 'lib';
use A;
use lib 'other-lib';
use B; # which uses A

A will be precompiled into lib/.precomp. B will be precompiled into other-lib
against A. So everyone should have been happy. But we tried to be extra
careful to not load a precomp file that was compiled when not the whole
repo-chain was available. We therefore included the identity of the whole repo
chain in a module's precomp id. So The A our B is precompiled against had a
different precomp id (due to the different repo-chain) than the A we loaded in
the main program. This caused the collision.

Since we now have the means to safely access precomp stores of the whole repo
chain, we don't need this extra carefulness anymore. Indeed it's actually
harmful.