Eliminate locking in favor of a work queue.
The approach taken so far is vulnerable to deadlock, since there can
be unfortunate timing interactions between the up-stream management
of the subscription and downstream flow of values. This changes puts
a work queue in place instead, which successfully avoids the deadlock
and in theory should give lower overhead, less blocked threads, and
hopefully better CPU cache behavior. Regresses a couple of tests,
however.