Remove a seemingly bogus $control.done.
Just because we received all the values does not mean we have sent
them on yet, nor that we will any time soon. If the limit is zero,
then a future control message might raise it and allow us to send
out the values, but if we already did `.done` on $control these
can never reach us.

It seems that this has always been wrong, however only recent work
addressing other issues has exposed this problem. It's a bit of a
mystery how this ever actually worked before.