Improve message on (1,2,3).invert to mention Pair
Binding of the map routine internally now requires list elements to bind
as Pair.  Fixes RT #128860, insofar as it improves the error message.

(The alternate approach of inserting a CATCH into the map closure could in
theory produce an even better message, but it appeared to slow things down
more than the Pair binding approach.)

Also, we now propagate is-lazy correctly through .invert and .antipairs.
Note that we cannot make the degenerate case of .invert the same as .antipairs,
because .invert must expand list values into multiple antipairs.  In any
case, .antipairs is 5-10 times faster since it doesn't have to check for
such expansion.