Fix -Ilib and friends not overriding installed dependencies
When checking the dependencies of a precomp file, we did not take development
version of modules in FileSystem repositories into account. I.e. with A
depending on B putting a new version of B into the lib directory and running
perl6 -Ilib -MA you would still get the precompiled version of A that uses the
installed version of B.

Fixed by having the identity of a FileSystem repository reflect its complete
content. So by changing the content we detect a change in the repository chain
and know to re-resolve dependencies.