Make importing globals from loaded modules lexical
Right now if A depends on C:auth<foo> and B depends on C:auth<bar> you
cannot use A and B in the same program. Fix this by having not only
exported symbols be lexical but also the globals contained in a module.

So only A will even have a C that's bound to C:auth<foo>. This means
that you have to "use" modules in every scope that accesses those names,
which is a good thing to do anyway.

This will break some existing code. See
http://rakudo.org/2016/12/17/lexical-module-loading/
for details.