Fix for RT #129044
If the Regex candidate needs to use .ACCEPTS, then it basically is
the same as the Mu candidate.  However, since we also have a Callable
candidate, and Regex.isa(Callable), that doesn't work.  So we abstract
the Mu candidate logic into two private methods, and call these from
both the Regex as well as the Mu candidate.

Also micro-optimize the :end versions, because assignment to $_ is
not needed for every iteration.