Initial support for error reporting plugins
This attempts to address RT #128983 .  Whenever an error is to be
reported, the RAKUDO_EXCEPTIONS_HANDLER environment variable is
checked: if it contains something, it is considered to be the name
of a class extending the Exceptions:: class (so "JSON" would refer
to the Exceptions::JSON class).  If such a class exists, then the
method "process" is called in that class, with the Exception object
as the parameter.  That method is than supposed to completely handle
the error reporting.

This patch also adds an Exceptions::JSON class that outputs the
exception information as JSON on STDERR.