First steps on fixing string async socket reads.
This uses the new VM-backed streaming decode API, which will allow for
much better error handling, choosing the encoding, and in the future
provide support for userspace decoders to be plugged in. This nearly
works, with a caveat: we run into a phaser scoping bug with the LAST
block that will need some attention. (Also, the streaming decode stuff
is not yet available for the JVM backend.)