Calling unary on reduce needs arity, not count
For a reduction with a single argument, we want to look to see if the operator
in question has a unary interpretation, and call that.  Unfortunately, there
are functions that can handle a unary but whose count is larger than 1,
so would not be called.  Now we check .arity instead, which can still be 1
(or 0) for such functions.  (We don't test for exactly 1 since most functions
with .arity of 0 have a .count greater than 0, so allow one arg.)