We need to be smart about terminating backtraces.
The sprintf implementation is done in NQP and parsed using a grammar.
On an error with the format or processing the format, we panic, which
is fine. However, panic is in NQPHLL. We used detection of NQPHLL in
order to trim the bottommost frames from stack traces (which are the
HLL::Compiler startup infrastructure, which is not useful to show).
Unfortuantely, this led to us trimming all the useful userspace
lines away from the backtrace of sprintf also, meaning the user had
no context for where the error came from. Sad! Also fixed it so that
panic does not show up in the backtrace, which apparently made some
users panic too. :-)