Add a .Map coercer for object hashes
Not sure this is the utlimate right fix.  But at this point in time,
we do not have object Map's.  So we cannot coerce an object Hash to
an object Map, so we need to coerce to a normal Map.  And that means
that the objects will become stringified in the Map.

This solution will mostly do the right thing.  On the other hand,
there doesn't seem to be a reason to do this (from a performance point
of view), because it will involve a lot of copying of data.