No longer read any JSON META data when loading modules
The only bit we needed from the META data was the id of the source file
for a module. Just add this to the short-name lookup file and take it
from there.
As the CompUnit contains a reference to the Distribution object, we now
give it a lazily initializing one that only reads the META data in the
rare case that someone actually accesses it.