Remove "this is only some changes" para from template
At least for now... Our changelogs are short enough to be included in entirety,
and there are usually just a couple of items that are omited. Including
all the items also makes generation of release announcements entirely
automatic, without requiring human intervention or extra infrastructure to
filter out items of small-interest.