Fix the x operator for large count values
There are two changes in this commit. The first is a fix in the initial
check that the count is greater than 0. The change in 586f784 meant that
large count values could overflow the native int size and become negative
numbers, thus not correctly passing the check and instead returning an
empty string.

However, before that change large values would instead overflow when
passed to nqp::x, which actually performs the string repetition. So the
second change is to check if the value passed in is larger than an int.
If so, use the max size of an int instead. The nqp::x op has an arbitrary
limit of 1073741824 (2**30), so it will error out, but if any changes to
that limit in NQP are made later, they should get picked up automatically.

See http://irclog.perlgeek.de/perl6-dev/2016-09-04#i_13148294 and
http://irclog.perlgeek.de/perl6-dev/2016-09-04#i_13148630 for some
discussion.

Fixes RT #128035