Update precomp unit cache after precompiling an outdated unit
We try to load a precomp file, discover that it's outdated and precompile it
again. When we then try again to load it, we found the old version in the cache
and happily used that again. Fix by updating the cache when storing the new
version.
Thanks to TheLemonMan++ fo the report.