Speed up loading deps by only running .modified checks when needed
$*REPO.id represents the contents of all repositories. We already check if
that changed since we precompiled a module. If it is still the same, there
should be no reason to check the .modified of the source file. If it changed,
we need to check it as a change in our source file could be exactly what
caused the different checksum.
Note that this relies on the immutability of installed distros which we
haven't enforced so far.
Brings GTK::Simple's load time back down to 227ms.