Fix NativeCall CArray hanging when created from empty list
The issue was we were calling nextsame when @values was empty,
but the nextsame candidate was the one with a slurpy, that
called the @values candidate again, resulting in a loop.

Instead of having a separate multi for empty argument case, we
return its value when the @values is empty, resolving the infiniloop.

Fixes RT#129256: https://rt.perl.org/Ticket/Display.html?id=129256