Make Cool.trans work the same as Str.trans
Cool.trans coerces the invovant to Str and then directly calls
Str.trans. The way this is done causes adverbs, if specified, to
be thrown away. We must ensure that the these parameters are sent
to Str.trans and not being thrown away. A couple of changes
suggested by Zoffix++ takes care of this problem.