Fixates borg and mystery at block start
Fixes RT #129290.

The way this used to work was complete blocks would set $*BORG<block>,
which missing_block would use to provide error information.  The problem
with this is if you start parsing a block (let's call it A), and create
a complete block (named B) within A.  If there is no closing brace to
terminate A, missing_block will be called, but the block context set
up for B will be used to explain the error because it was the last
block to successfully parse.

This change alters all of the callsites for missing_block so that we
capture the values of $*BORG (and $*MYSTERY, which is also used for
explaining errors) right before we try to match the opening brace, and
pass them manually when and if we call missing_lock.  If something is
swallowing the block, it would have already done so at this point.