Fix for RT #129299
The List.iterator assumed that a List could never have a "hole" in it.
This allowed the iterator to not have to check the index against the
number of elements in the list.  I have now reinstated that check,
albeit only when it finds what could be a hole.  If the index is then
still less than the number of elems, we return the hole and continue.

This adds an extra test for *each* List.iterator.  Going to think about
another way to fix this if this shows up in reduced performance somehow.