Fix basic native attributive parameter binding.
This handles the case where the native attribute appears as either a
named or positional parameter inside of a signature of a method in a
class, where the signature is simple enough to be lowered (so, does
not contain a sub-signature). Complex signatures and methods in roles
will need additional, likely more involved, fixes.