Allow an Enum in an array shape declaration
As suggested in https://gfldex.wordpress.com/2016/09/22/keys-are-optional/

It will take the number of elements of the Enum as the shape of the
array.  It is provided here as a proof of concept.  Not sure this should
be part of v6.c.1 or v6.d , or that it can just live here and now.

Allowing support for this seems like a good thing, as it enforces compile
time checks on validity of array indexes.  Which should allow us to
optimize these checks away later at runtime.