Fix GLOBAL symbol clash re-compiling due to a repo change
After we detected that we need to re-compile a module because one of its
dependencies changed, we sometimes turned out to try loading the old
version again due to a stale precomp unit cache. However the module
would load just fine on the next try.

The reason was that we invalidated the cache of a different precomp
store object than what we used for loading precomp files as the
repository implementations created different objects for the precomp
store list (used for loading) and the accessors of their own store (used
for storing).