Revert "Fix Hash.classify-list for mixed-level classifications"
This reverts commit b4a4b604945772a828e9edb0fa8b1cb295e8bf7a.

Mixed-level classifications cause issues when values at one level
are same as keys for deeper levels and speculation explicitly says
all lists returned from mappers for multi-level classifications must
be of same length.

Instead, logic needs to be added to check that all levels are the same,
so we do not produce confusing error messages when they aren't the same.