Fix for RT #129321
Under some circumstances duckmap could recurse indefinitely. The problem
was that if the expression returned undef for an argument the code would
gladly take the same argument, which caused the problem, and call
duckmap again leading to a hang or OOM condition. After a discussion with
with TimToady++ in #perl6 he came up with a solution to the problem which
is what this commit contains.