Fix Int.new for large values
Currently Int.new in range 2⁶³-2⁶⁴ gives incorrect negative results
and gives unboxing errors for larger values.

Fix it by coercing the value to an .Int, so we can handle a larger range of
possible accepted values, and then cloning the value, so we get a new object,
that is decoupled from any cached numeric constants.