Port !cursor_more and !cursor_next from nqp
Wanted to change the interface to !cursor_more specifically in nqp
so that it wouldn't need to take a hash and do lookups.  Especially
since it only checked for <ex> to call !cursor_next and <ov> to see
whether an overlap needed to be handled.  The new Cursor.CURSOR_MORE
now only checks for overlap: it is expected that the new .CURSOR_NEXT
is called directly if exhaustive is set.