Convert more BOOTSTRAPATTRs to Attributes
This is a second pass, changing individual instances of BOOTSTRAPATTR to
Attributes even if not all for a class could be converted. Additionally,
this includes a change to scalar_attr() to permit creating Attributes
without setting :auto_viv_container.

The change to Code's $!compstuff allows the JVM build to compile;
however, the install still fails with 'Flattening named argument must
have VMHash REPR'.