Add Iterator.skip-at-least-pull-one
Skip at least N values of the iterator, and return the result of the
next pull-one.

Initially I thought about adding this as a method to Rakudo::Internals,
but it seemed more logical to add it here.  Although this may have
adverse effects, as now every class that does Iterator will have a copy
of this method?