Fix Cool.IO coercion
Currently, calling .IO on a type object attempts to create a new IO::Path
using it, so we end up with gut-referencing warnings. This is especially
harmful with coersions on optional args, as sub (IO(Str) :$foo) {}(), for
example will emit undef-in-Str warnings.

Do the coersion only for :D params. For :U, return an IO::Path type object.

Note: not sure what the purpose of |c was. IMO, its presents allows for users
to call the method with some args, expecting the args to do something.
I removed it, to prevent such silent calls with args that do nothing.
Stresstest passes.