Introduce Rakudo::Internals.SeqUsingIndexIterator
Create a Seq from a Seq and an iterator supplying index values.
Originally created as part of the Str.match refactor work, but
turned out to be not needed as part of that.  So for now, this
will sit here until we start looking at slicing of (multidim)
arrays again.  Or be removed should it turn out not to be needed
after all.