Make is-approx DWIM when no tolerances are provided
When current version was designed, the default was chosen to abs-tol of 1-e5, with
reasoning behind that it would be easy for users to understand it (as compared with
rel-tol or some DWIM mechanism). While that goal was accomplished, when using the
function in real-world application, the abs-tol default proves to be nearly useless.
It's too small for large numbers and too large for small numbers, forcing the user to
specify manually. Moreover, rel-tol can't be used when $expected is zero, so the user
constantly has to think about which tolerances to plug into the function.

This commit restores the DWIMmy behaviour of old is_approx for the default case where
no tolerances are given: if $expected.abs() is smaller than 1e-6, use absolute
tolerance of 1e-5; else, use relative tolerance of 1e-6.

Both the existing Rakudo tests for this function and the entire stresstest passes with
these changes, but that's mostly due to undertesting (or in roast's case, no testing).
I looked at ecosystem usages and the function isn't used frequently. Both of those
suggest to me the impact of this change would be minimal, if not entirely non-existent.