Introduce Iterator.skip-one
A method that can be overriden if you're writing an iterator that can
be significantly cheaper when skipping a generated value.  Use case,
the :10nth parameter with Str.match (aka, give me only the 10th match).
By default, it is just a slightly more expensive .pull-one.

Also use let Iterator.skip-at-least-pull-one use skip-one, as well
as in Rakudo::Internals.SeqSkipNFromIterator .