Make QuantHash:U AT-KEY autovivify to QuantHash object, not Hash
Currently, when attempting to do AT-KEY operations on Setties and Baggies,
the autovivification is handled by Any:U.AT-KEY, which defaults to creating
a Hash object. This causes issues with my BagHash $b; $b<foo>++, since an
attempt to assign a Hash to BagHash variable results in an exception.

Fix by adding AT-KEY for QuantHash:U; a role that Setties and Baggies do
that will create an appropriate object instead of a Hash.