Fix acosh for large negative numbers
The IEEE 754-2008[^1], Section 9.2, says acosh for values < 1 is a NaN.
We do that for numbers with small maginitude and -∞ itself, but for
large negative numbers (e.g. -1e100 and -1e200) we end up returning ±∞

Fix by detecting out-of-domain input and returning NaN without performing
any calculations. This commit fixes Num and num versions, but not Complex.
I need to do further research on complex numbers to know how to fix them.

[1] http://www.csee.umbc.edu/~tsimo1/CMSC455/IEEE-754-2008.pdf