Rakudo::Internals.IterateFromIterators
Create iterator from two iterators: one with source values, and one with
index values.  Cache the source elements that are requested by the index
values.  Indexes do not need to be in any order.  If there is no value
for a given index (because it's not in the cache and the source values
iterator is exhausted), return Nil.  Optionally call an out-of-bounds
block, e.g. to allow warnings or setting flags.

Remove SeqUsingIndexIterator, which was the inspiration.

This iterator will be used for Str.match(/./, :nth(2,6,4))