Completely re-imagine Str.match internals
Simple matches without any modifying adverbs is about 1.5x faster.
Complicated setups like with :nth(1,3..*), :3x are about 14x faster.
Other mixes of parameters are generally in the 2x to 3x faster range.

Two semantic changes: (see RT #129945): the family of :nth (as in :st,
:nd, :rd and :th) are now lazy and will return whatever they can find.
Before all matches would be cached to find out if enough matches were
found and none would be returned if not enough were found.  Please
use an additional :x for that if you want that behaviour.  Also, when
using .nth with an iterator using non-monotonically increasing index
values, it will now die, instead of silently not producing the Matchi
object associated with the non-monotonically increasing index value.

How this was done:
- use MMD for handling single adverb calls faster
- rewrite all (private) methods using nqp
- use iterators instead of gather/take combo's
- use new "skip-one" iterator methods wherever possible
- don't use signatures for multiple named variables
- pass $/ to be set around as a parameter