Optimize Str.match a bit more
Basically found a way to create a configurable iterator that doesn't
take more than the previous approach using several different iterators.
This easily allowed a much cheaper iterator that just returns Cursor
objects to be made, and used specifically in the :nth(*) and :nth(*-1)
cases, which are now about 2x as fast (for a regexp with 10 matches)
because only on the last pulled value a .MATCH needs to be done.

Also built the infrastructure to support iterators returning Str and
Pairs with from/length information, for later usage in .comb and .subst.