Don't use indexes
Again, if we're going to define "constants", they might as well be the
actual thing.  So get rid of a level of indirection, which gives us
about 1%.  Not a lot, but for such a vital part of the language worth
it.  And it should also improve readability for a change  :-)