Fix our scoped nested package swallowed by lexically scoped parent
In "use Foo; class Foo::Bar { }" Foo::Bar would be created as child of
the lexically imported Foo class and was thus missing from the globals.
Previous fixes only handled the case when Foo was a stub package, not
when it was a full class.
Remove the special case and fix the follow up bug that we now got a
duplicate symbol when trying to lexically override classes in the
RESTRICTED setting. There's no need to upgrade the main package to a
global if we only want to create a lexically scoped child anyway.