Bump for MoarVM with nativecall/GC fixes.
With this fix, we make sure to GC-block the current thread before we
make a nativecall, and unblock it afterwards (taking care with the
handling of callbacks). This fixes an issue where calling a long
running bit of native code on one thread could make the others all
hang at the next GC until the native code completed.