Revert "Fix off by one errors in Failure messages"
It is fully legal to provide a start position *after* the string.
This also applies, to e.g. substr: "foo".substr(3) is the empty
string, "foo".substr(4) returns a Failure.

The only thing that's not supported, is large Ints, because we
internally only work with native ints.  And if we really want to
return information about what a valid position to start looking
at, we should also take the number of chars of the needle into
account.

This reverts commit c59ab1888db616126da217a436670e06f9da3c23.