Introduce R:I:ShapeIterator
This is most of the logic of the previously anonymous class inside
R:I:ShapeIndexIterator method abstracted, so that we can use it in
other places as well.  Reform the ShapeIndexIterator to use this
role in its anonymous class.  Performance degradation is minimal.
This will allow us to use this logic much more easily for methods
such as values, kv, pairs, antipairs, as well allow use to more
easily create custom versions of these for 2/3dimmed arrays.