Handle negative Mixy weights in ∪ operator
Lookup of a non-existant key in a Mixy gives a zero, which when `max`ed against
a key with a negative weight would lead to that key being removed from the final union,
making the operator trim all the negative weights for keys present in just one Mixy.

Add a check for whether a key exists: if it does, take a `max` value, if it doesn't,
just use the value of the key from other Mixy as is.

ab5tract++ for submitting code that revealed this bug.