Mix in shapedness in the class, rather than instance
This means we only have 1 global deopt for creating a shaped array
of a given number of dimensions and type.

- fix .of on shaped typed arrays
- fix the .perl of shaped typed arrays

This gives a 10% gain on a simple "for ^10000 { my @a[10] }" loop,
but for larger programs it should be much more because it would no
longer do a global deopt for every iteration.

Now we only need to figure out how to do this de-opt only at
compile time, instead of every time at run-time.