Bump for latest MoarVM.
This includes fixes to a SEGV in multi-threaded programs where strings
internally represented as strands were used in various ways (typically
as hash keys). As a bonus, hashes now use 4 (32-bit) or 8 (64-bit)
bytes less memory per item stored, so a hash of 1000 keys will be
up to 8KB smaller.