Make `parse` try a bit harder to get a match.
While a Grammar built out of regexes isn't a particularly wise idea,
if we do have a `regex TOP` (or the `rule` we specify to `parse` is a
`regex`) then we should at leave give it a chance to do the required
backtracking to produce a result that eats the whole string. It costs
nothing for non-regex grammars to add this.