Unify accepted .base() arguments
Currently second argument to base can be a Whatever, but only if the
invocant is a Rational. This creates a special case for the method, and
the user has to watch out what type of a Real the invocant is, before
using the Whatever precision.

Unify the method's behaviour by providing degenerate cases of Whatever
precision for the rest of the Reals where it acts as if precision was not specified.