Implement attributive parameter binding on the JVM.
As noted in the patch itself, there seem to be cases where we lose the
flattened STables in the REPRData for a given STable. As of now this patch just
ignores that and tries to DTRT regardless, which means falling back to the
previous behavior of simply assuming a boxed Attribute.