Fix Match pretending to be a value type when it is not
Comparing the identity of Match objects wrongly resulted in them being
treated as equals. Happened because Match is a subclass of Capture which
is a value type and therefor provides a WHICH method. Matches are most
certainly no value types as they are mutable. So we circumvent Capture's
WHICH method.