Show original program name in MAIN_HELPER instead of 'command'
Installed scripts are stored as resources in a
CompUnit::Repository::Installation and called via a wrapper script
stored in path.
$*PROGRAM-NAME contains the path to the resource while the user should
only have to deal with the wrapper script. So use an environment
variable to override $*PROGRAM-NAME in that case and show the wrapper
script's path in MAIN_HELPER error messages instead.