Use Supplier::Preserving in Proc::Async.
I recently got bitten by this, when writing code that obtained stdout
and stderr, put them in an object, did .start on the process, and then
returned the object. The calling code then tapped the two supplies.
This resulted in occasional hangs, since sometimes the output was
produced sooner than the object's user tapped. We could argue that you
Just Shouldn't Do This, but it's likely plenty of people will fall
into this trap over time, and it's just how most supplies work
anyway (you don't lose incoming packets on a Socket by not tapping
soon enough for example). So, whack a `::Preserving` on it.