Fix Iterator.flat occasionally skipping inner iterables
Fixes RT#130160: #https://rt.perl.org/Ticket/Display.html?id=130160
lucasb++ for golfing the bug.

The issue observed in the ticket occurs when a combination of .pull-one
and .push-all method is called on the iterator of the Seq's returned by
.flat(). Current implementation of .push-all on that iterator does not
anticipate that we *already* may be inside a nested iterable, so when
.pull-one stops instead a nested iterable, the .push-all call will
continue with whatever is *after* that iterable.

Fix by deleting the push-all implementation and using the default one
provided by Iterator. I've done rough performance check by running
this a few times: `for ^1000 { ^10_000 .flat.eager }; say now - INIT now;`
and the results point to a the new way being 1.0000019387x faster, so
there's no apparent performance loss.