Fix infinit loop in .subst with no arguments
If .subst method is called without arguments, the matching multi is one
provided by Cool. It coerces the arg to Stringy and calls .subst method
on it, passing args via a Capture. However, if that Capture doesn't satisfy
Stringy candidates, we get the Cool candidate again, and the loop repeats.

Fix by making Cool.subst an only method and moving the proto to Str.subst,
making only Str's candidates available.

ugexe++ for spotting:
https://irclog.perlgeek.de/perl6-dev/2016-11-26#i_13630780