Teach Str.Num to handle negative zeros
While val will return[^1] negative zeros when the string contains a num,
it won't do so for other types. Since we do have the sign information
available in the string itself and the method requests a Num, attach the
correct sign to zeros, even if the string does not contain a num.

Fixes RT#128897: https://rt.perl.org/Ticket/Display.html?id=128897

[1] https://github.com/rakudo/rakudo/commit/e2587cd012