Refactor and fix some races in supply/whenever.
This moves various mutations into the SupplyState object, where we can
reliably enforce mutual exclusion upon them. This of note eliminates
some races relating to the active-taps hash, which led to a number of
different occasional crashes. (It turns out that some of the mutations
need to happen outside of the regime for running user code.)