Fix .first on Numerics
The logic in the method uses `+self` to get number of elements in
the list, however, `self` is not guaranteed to be a list and may be
a Numeric type, in which case we get an incorrect number.

Fix by using `self.elems` instead.

Bug find: https://irclog.perlgeek.de/perl6/2016-12-08#i_13705826