Make .match(:g) return empty List on match failure
As a side-effect of Great .match Refactor[^1], .match(:g) on failure
to match was changed from returning an empty List to returning
an Empty (which is a slip). As a consequence, the idiom of
nqp::p6store( nqp::call(&infix:<,>, $/), $_) (where $/ is the slip), in
Perl6/Actions that handle S:g/// resulted in an empty list, rather than
the actual original string, which is the cause of the bug in the ticket[^3]

During the G.mR, the .comb on failure was also changed to return an Empty,
instead of empty Seq. This commit affects that behaviour as well,
returning an empty List instead.

[1] https://github.com/rakudo/rakudo/commit/b7201a8f223
[2] https://github.com/rakudo/rakudo/blob/c4a60110ef/src/Perl6/Actions.nqp#L7796-L7803
[3] https://rt.perl.org/Ticket/Display.html?id=130289