Make infix:<..> on Nums 13x faster + fix native num NaN ===
And anything else that uses === on Nums should be faster too.
jnthn++ for finding the cause of slowdown.

The presence of &infix:<===>(NaN, NaN) candidate prevents
caching[1], which causes massive slowdown in infix:<..> on Nums.
The candidate also does not handle native nums as they go to a
narrower candidate, so NaN === NaN is broken too.

Fix by removing (NaN, NaN) candidate and adding a check for NaNs
into other candidates. Any num but NaN is less-or-equal to positive
infinity, so use that as an `nqp::isnan()` check.

Fixes RT#130190: https://rt.perl.org/Ticket/Display.html?id=130190

[1] https://irclog.perlgeek.de/perl6/2016-12-09#i_13713052