Fix imported nested enum colliding with symbol from outer scope
We also search outer scopes for declared packages, in case we import a sub-
package. But when we actually want to lexically overwrite the symbol from the
outer scope, we would wrongfully detect a collision. Fix by allowing for
overwriting the symbol from the outer lexical scope after detecting that we
cannot just merge packages.
Reproducible by having A.pm: enum A::B <Date>;
perl6 -I. -e 'use A'