Prevent Mu.clone from incorrectly marking all attrs as inited
Fixes RT#130271: https://rt.perl.org/Ticket/Display.html?id=130271

Using nqp::getattr() or nqp::bindattr() on an attribute makes
nqp::attrinited return true. This made cloned Baggies lose their
.WHICH value, since its method uses nqp::attrinted to decide whether
to give its $!WHICH attribute. And in the ticket we have a cloned
Baggy, since the infix:<but> operator clones under the hood. And
this is the reason for the error in the ticket, since the .WHICH value
is unboxed into a native str at one point.

Fix by checking whether an attribute is nqp::attrinted before to
decide whether to call getattr/bindattr.