Fix crash on Rat Ranges constructed with Range ops
When using range ops, we clone the Range. However, the code
makes a mistaken assumption that Mu.clone can tweedle anything,
but only things with accessors can be passed as tweedles. As a
result, the private $!is-int attribute wasn't being set, leading
to an explosion when a Rat range was being treated as an Int Range.

Fix by setting $!is-int attr in the clone ourselves.

Fixes RT#128887: https://rt.perl.org/Ticket/Display.html?id=128887