Fix andthen → orelse chaining
Fixes RT#130034 https://rt.perl.org/Ticket/Display.html?id=130034

Per TimToady[^1], `andthen` returning an Empty is by design, and the
behaviour  seen in the ticket is due to the RHS `orelse` thunk not being
executed but returned as is. Since `orelse` isn't defined as executing any
Callables on LHS, we need to somehow figure out when we get the Empty
(which vanishes in our parameters when we receive it). So I added a check
for whether we got just 1 arg (our RHS thunk) and then eval it if it's
a Callable. So the behaviour is that orelse with 1 arg evals its callables.

Not overly happy with this fix, but it fixes the bug and passes stresstest.

[1] https://irclog.perlgeek.de/perl6-dev/2016-11-11#i_13550043